Add files via upload
The program will prompt you with the name of the item whose per unit price you want to find. The program then asks you how expansive of a search you want to conduct, then using APIs, BERTs specifically a Robustly Optimized BERT, and then it sifts through the results and gives you all results with a certainty score of 0.70 or greater.