Add files via upload
This is a voice assistant that I have made that can record and create logs create summaries of those logs and can sift through the internet for specific questions that are not specific commands. I plan to work in the future on making better APIs for web searching for it and use newer NLP programs to improve comprehension of the analyzed websites.